[AOTI] Store OpOverload in ir.ExternKernel
Summary: Currently the logics for filling the default value for optional arguments are scattered in several places. By storing OpOverload in the base ExternKernel class, we can simplify codegen_kwargs, and this is a preparation step for enabling the torchgen-ed C shim. The default value filling logic for FallbackKernel can also be simplified, but that can come later.

ghstack-source-id: aade8b926749390c888004f8847108cc19cd7f0f
Pull Request resolved: https://github.com/pytorch/pytorch/pull/120629